Move context props into single file
