+++ packages/react-day-picker/src/contexts/DayPicker/DayPickerContext.ts
@@ -1,6 +1,38 @@
 import * as React from 'react';
 
-import { DayPickerContextValue } from './DayPickerContextValue';
+import {
+  CaptionLayout,
+  ClassNames,
+  Components,
+  DayPickerProps,
+  Formatters,
+  Labels,
+  ModifierClassNames,
+  Modifiers,
+  Styles
+} from '../../types';
+
+/** Represent the value of the [[DayPickerContext]]. */
+export type DayPickerContextValue = DayPickerProps & {
+  captionLayout: CaptionLayout;
+  classNames: Required<ClassNames>;
+  components: Components;
+  formatters: Formatters;
+  labels: Labels;
+  locale: Locale;
+  modifierClassNames: ModifierClassNames;
+  modifierPrefix: string;
+  modifiers: Modifiers;
+  numberOfMonths: number;
+  styles: Styles;
+  today: Date;
+
+  // Internally we handle only fromDate/Todate
+  toYear?: never;
+  fromYear?: never;
+  toMonth?: never;
+  fromMonth?: never;
+};
 
 /**
  * The DayPicker context shares the props passed to DayPicker within the
