--- packages/react-day-picker/src/contexts/Focus/FocusContextValue.ts
@@ -1,19 +1,0 @@
-/** Represents the value of the [[NavigationContext]]. */
-export type FocusContextValue = [
-  /** The day currently focused */
-  focusedDay: Date | undefined,
-  setters: {
-    /** Focus the specified day. */
-    focus: (day: Date) => void;
-    /** Blur the focused day */
-    blur: () => void;
-    /** Focus the day after the focused day. */
-    focusDayAfter: () => void;
-    /** Focus the day before the focused day. */
-    focusDayBefore: () => void;
-    /** Focus the day in the week before the focused day. */
-    focusWeekBeforeDay: () => void;
-    /** Focus the day in the week after the focused day. */
-    focusWeekAfterDay: () => void;
-  }
-];
