+++ packages/react-day-picker/src/contexts/Focus/index.ts
@@ -1,4 +1,3 @@
 export * from './FocusContext';
-export * from './FocusContextValue';
 export * from './FocusProvider';
 export * from './useFocus';
