+++ packages/react-day-picker/src/contexts/Focus/useFocus.ts
@@ -1,7 +1,6 @@
 import * as React from 'react';
 
-import { FocusContext } from './FocusContext';
-import { FocusContextValue } from './FocusContextValue';
+import { FocusContext, FocusContextValue } from './FocusContext';
 
 /** Hook to access the [[FocusContext]]. */
 export function useFocus(): FocusContextValue {
