+++ packages/react-day-picker/src/contexts/Navigation/index.ts
@@ -1,4 +1,3 @@
 export * from './NavigationContext';
-export * from './NavigationContextValue';
 export * from './NavigationProvider';
 export * from './useNavigation';
