--- packages/react-day-picker/src/contexts/SelectMultiple/SelectMultipleModifiers.tsx
@@ -1,4 +1,0 @@
-import { Modifiers } from '../../types';
-
-/** Represent the modifiers that are changed by the multiple selection. */
-export type SelectMultipleModifiers = Pick<Modifiers, 'selected' | 'disabled'>;
