+++ packages/react-day-picker/src/contexts/SelectMultiple/index.ts
@@ -1,5 +1,3 @@
 export * from './SelectMultipleContext';
-export * from './SelectMultipleContextValue';
-export * from './SelectMultipleModifiers';
 export * from './SelectMultipleProvider';
 export * from './useSelectMultiple';
