--- packages/react-day-picker/src/contexts/SelectRange/SelectRangeModifiers.tsx
@@ -1,7 +1,0 @@
-import { Modifiers } from '../../types';
-
-/** Represent the modifiers that are changed by the range selection. */
-export type SelectRangeModifiers = Pick<
-  Modifiers,
-  'selected' | 'range_start' | 'range_end' | 'range_middle' | 'disabled'
->;
