+++ packages/react-day-picker/src/contexts/SelectRange/SelectRangeProvider.tsx
@@ -9,8 +9,7 @@ import {
   isDayPickerRange
 } from '../../types';
 
-import { SelectRangeContext } from './SelectRangeContext';
-import { SelectRangeModifiers } from './SelectRangeModifiers';
+import { SelectRangeContext, SelectRangeModifiers } from './SelectRangeContext';
 import { addToRange } from './utils/addToRange';
 import { useControllablePropState } from '../../hooks/useControllablePropState';
