+++ packages/react-day-picker/src/contexts/SelectRange/index.ts
@@ -1,5 +1,3 @@
 export * from './SelectRangeContext';
-export * from './SelectRangeContextValue';
-export * from './SelectRangeModifiers';
 export * from './SelectRangeProvider';
 export * from './useSelectRange';
