+++ packages/react-day-picker/src/contexts/SelectSingle/SelectSingleContext.tsx
@@ -1,6 +1,22 @@
 import * as React from 'react';
 
-import { SelectSingleContextValue } from './SelectSingleContextValue';
+import { DayClickEventHandler } from '../../types';
+
+import { Modifiers } from '../../types';
+
+/** Represent the modifiers that are changed by the single selection. */
+
+export type SelectSingleModifiers = Pick<Modifiers, 'selected'>;
+
+/** Represents the value of a [[SelectSingleContext]]. */
+export interface SelectSingleContextValue {
+  /** The day that has been selected. */
+  selected: Date | undefined;
+  /** The modifiers for the corresponding selection. */
+  modifiers: SelectSingleModifiers;
+  /** Event handler to attach to the day button to enable the single select. */
+  handleDayClick: DayClickEventHandler;
+}
 
 /**
  * The SelectSingle context shares details about the selected days when in
