--- packages/react-day-picker/src/contexts/SelectSingle/SelectSingleModifiers.tsx
@@ -1,5 +1,0 @@
-import { Modifiers } from '../../types';
-
-/** Represent the modifiers that are changed by the single selection. */
-
-export type SelectSingleModifiers = Pick<Modifiers, 'selected'>;
