+++ packages/react-day-picker/src/contexts/SelectSingle/index.ts
@@ -1,5 +1,3 @@
 export * from './SelectSingleContext';
-export * from './SelectSingleContextValue';
-export * from './SelectSingleModifiers';
 export * from './SelectSingleProvider';
 export * from './useSelectSingle';
